# feat: optimistically compile go modules while extracting schema